+++ social/apps/flask_app/peewee/models.py
@@ -1,1 +1,1 @@
-fro social_flask_peewee.models import FlaskStorage, init_social
+from social_flask_peewee.models import FlaskStorage, init_social